+++ eventkit_cloud/ui/static/ui/app/tests/actions/datacartActions.spec.js
@@ -4,13 +4,41 @@ import thunk from 'redux-thunk';
 import axios from 'axios';
 import sinon from 'sinon';
 import MockAdapter from 'axios-mock-adapter';
-import types from '../../actions/actionTypes';
 import * as actions from '../../actions/datacartActions';
 
 const middlewares = [thunk];
 const mockStore = configureMockStore(middlewares);
 
 describe('export actions', () => {
+    const expectedRuns = [
+        {
+            uid: '123',
+            url: 'http://cloud.eventkit.test/api/runs/123',
+            started_at: '2017-03-10T15:52:35.637331Z',
+            finished_at: '2017-03-10T15:52:39.837Z',
+            duration: '0:00:04.199825',
+            user: 'admin',
+            status: 'COMPLETED',
+            visibility: 'PRIVATE',
+            job: {
+                uid: '123',
+                name: 'Test1',
+                event: 'Test1 event',
+                description: 'Test1 description',
+                url: 'http://cloud.eventkit.test/api/jobs/123',
+                extent: {},
+                selection: '',
+                permissions: {
+                    groups: {},
+                    users: {},
+                },
+            },
+            provider_tasks: [],
+            zipfile_url: 'http://cloud.eventkit.test/downloads/123/test.zip',
+            expiration: '2017-03-24T15:52:35.637258Z',
+        },
+    ];
+
     const geojson = {
         type: 'FeatureCollection',
         features: [
@@ -110,8 +138,8 @@ describe('export actions', () => {
         mock.onPost('/api/jobs').reply(200, { uid: '123456789' });
 
         const expectedActions = [
-            { type: types.SUBMITTING_JOB },
-            { jobuid: '123456789', type: types.JOB_SUBMITTED_SUCCESS },
+            { type: actions.types.SUBMITTING_JOB },
+            { jobuid: '123456789', type: actions.types.JOB_SUBMITTED_SUCCESS },
         ];
 
         const store = mockStore({ jobSubmit: { jobuid: {} } });
@@ -127,41 +155,12 @@ describe('export actions', () => {
         mock.onPost('/api/jobs').reply(400, 'uh oh');
 
         const expectedActions = [
-            { type: types.SUBMITTING_JOB },
-            { type: types.JOB_SUBMITTED_ERROR, error: 'uh oh' },
+            { type: actions.types.SUBMITTING_JOB },
+            { type: actions.types.JOB_SUBMITTED_ERROR, error: 'uh oh' },
         ];
 
         const store = mockStore({});
         return store.dispatch(actions.submitJob(jobData))
-            .then(() => {
-                expect(store.getActions()).toEqual(expectedActions);
-            });
-    });
-
-    it('getProviders should return providers from the api', () => {
-        const mock = new MockAdapter(axios, { delayResponse: 1 });
-        mock.onGet('/api/providers').reply(200, ['my providers']);
-
-        const expectedActions = [
-            { type: types.GETTING_PROVIDERS },
-            { type: types.PROVIDERS_RECEIVED, providers: ['my providers'] },
-        ];
-        const store = mockStore({});
-        return store.dispatch(actions.getProviders())
-            .then(() => {
-                expect(store.getActions()).toEqual(expectedActions);
-            });
-    });
-
-    it('getProviders should handle errors', () => {
-        const mock = new MockAdapter(axios, { delayResponse: 1 });
-        mock.onGet('/api/providers').reply(400);
-
-        const expectedActions = [
-            { type: types.GETTING_PROVIDERS },
-        ];
-        const store = mockStore({});
-        return store.dispatch(actions.getProviders())
             .then(() => {
                 expect(store.getActions()).toEqual(expectedActions);
             });
@@ -172,8 +171,8 @@ describe('export actions', () => {
         mock.onGet('/api/formats').reply(200, ['my formats']);
 
         const expectedActions = [
-            { type: types.GETTING_FORMATS },
-            { type: types.FORMATS_RECEIVED, formats: ['my formats'] },
+            { type: actions.types.GETTING_FORMATS },
+            { type: actions.types.FORMATS_RECEIVED, formats: ['my formats'] },
         ];
         const store = mockStore({});
         return store.dispatch(actions.getFormats())
@@ -187,7 +186,7 @@ describe('export actions', () => {
         mock.onGet('/api/formats').reply(400);
 
         const expectedActions = [
-            { type: types.GETTING_FORMATS },
+            { type: actions.types.GETTING_FORMATS },
         ];
         const store = mockStore({});
         return store.dispatch(actions.getFormats())
@@ -211,8 +210,8 @@ describe('export actions', () => {
     it('DrawerTimeout closeDrawer should close drawer', () => {
         const timeout = new actions.DrawerTimeout();
         const expectedActions = [
-            { type: types.CLOSING_DRAWER },
-            { type: types.CLOSED_DRAWER },
+            { type: actions.types.CLOSING_DRAWER },
+            { type: actions.types.CLOSED_DRAWER },
         ];
 
         const store = mockStore({ drawer: 'open' });
@@ -226,8 +225,8 @@ describe('export actions', () => {
         const clearStub = sinon.stub(global, 'clearTimeout');
         const timeout = new actions.DrawerTimeout(null, 'yo');
         const expectedActions = [
-            { type: types.CLOSING_DRAWER },
-            { type: types.CLOSED_DRAWER },
+            { type: actions.types.CLOSING_DRAWER },
+            { type: actions.types.CLOSED_DRAWER },
         ];
 
         const store = mockStore({ drawer: 'open' });
@@ -242,8 +241,8 @@ describe('export actions', () => {
     it('DrawerTimeout openDrawer should open drawer', () => {
         const timeout = new actions.DrawerTimeout();
         const expectedActions = [
-            { type: types.OPENING_DRAWER },
-            { type: types.OPENED_DRAWER },
+            { type: actions.types.OPENING_DRAWER },
+            { type: actions.types.OPENED_DRAWER },
         ];
 
         const store = mockStore({ drawer: 'closed' });
@@ -257,8 +256,8 @@ describe('export actions', () => {
         const clearStub = sinon.stub(global, 'clearTimeout');
         const timeout = new actions.DrawerTimeout('yo', null);
         const expectedActions = [
-            { type: types.OPENING_DRAWER },
-            { type: types.OPENED_DRAWER },
+            { type: actions.types.OPENING_DRAWER },
+            { type: actions.types.OPENED_DRAWER },
         ];
 
         const store = mockStore({ drawer: 'closed' });
@@ -268,5 +267,87 @@ describe('export actions', () => {
                 expect(clearStub.calledOnce).toBe(true);
                 clearStub.restore();
             });
+    });
+
+    it('reRunExport should return a specific run from "api/runs"', () => {
+        const mock = new MockAdapter(axios, { delayResponse: 10 });
+        mock.onPost('/api/jobs/123456789/run').reply(200, expectedRuns);
+        const expectedActions = [
+            { type: actions.types.RERUNNING_EXPORT },
+            { type: actions.types.RERUN_EXPORT_SUCCESS, exportReRun: { data: expectedRuns } },
+        ];
+
+        const store = mockStore({ exportReRun: {} });
+
+        return store.dispatch(actions.rerunExport('123456789'))
+            .then(() => {
+                expect(store.getActions()).toEqual(expectedActions);
+            });
+    });
+
+    it('reRunExport should dispatch an error', () => {
+        const mock = new MockAdapter(axios, { delayResponse: 10 });
+        mock.onPost('/api/jobs/123/run').reply(400, 'oh no an error');
+        const expectedActions = [
+            { type: actions.types.RERUNNING_EXPORT },
+            { type: actions.types.RERUN_EXPORT_ERROR, error: 'oh no an error' },
+        ];
+
+        const store = mockStore({ exportReRun: {} });
+
+        return store.dispatch(actions.rerunExport('123'))
+            .then(() => {
+                expect(store.getActions()).toEqual(expectedActions);
+            });
+    });
+
+    it('clearReRunInfo should return type CLEAR_RERUN_INFO and no action', () => {
+        expect(actions.clearReRunInfo()).toEqual({
+            type: 'CLEAR_RERUN_INFO',
+        });
+    });
+
+    it('updateDataCartPermissions should dispatch a patch and update the published state on the job', () => {
+        const mock = new MockAdapter(axios, { delayResponse: 10 });
+
+        mock.onPatch('/api/jobs/123456789').reply(204);
+        const expectedActions = [
+            { type: actions.types.UPDATING_PERMISSION },
+            { type: actions.types.UPDATE_PERMISSION_SUCCESS },
+        ];
+
+        const store = mockStore({ updatePermission: {} });
+
+        return store.dispatch(actions.updateDataCartPermissions('123456789', {
+            permissions: {
+                value: 'SHARED',
+                groups: { group_one: 'READ' },
+                members: { admin: 'ADMIN' },
+            },
+        })).then(() => {
+            expect(store.getActions()).toEqual(expectedActions);
+        });
+    });
+
+    it('updateDataCartPermissions should dispatch an error', () => {
+        const mock = new MockAdapter(axios, { delayResponse: 10 });
+
+        mock.onPatch('/api/jobs/123').reply(400, 'oh no an error');
+        const expectedActions = [
+            { type: actions.types.UPDATING_PERMISSION },
+            { type: actions.types.UPDATE_PERMISSION_ERROR, error: 'oh no an error' },
+        ];
+
+        const store = mockStore({ updatePermission: {} });
+
+        return store.dispatch(actions.updateDataCartPermissions('123', {
+            permissions: {
+                value: 'SHARED',
+                groups: { group_one: 'READ' },
+                members: { admin: 'ADMIN' },
+            },
+        })).then(() => {
+            expect(store.getActions()).toEqual(expectedActions);
+        });
     });
 });
