+++ eventkit_cloud/ui/static/ui/app/tests/actions/licenseActions.spec.js
@@ -2,8 +2,7 @@ import configureMockStore from 'redux-mock-store';
 import thunk from 'redux-thunk';
 import axios from 'axios';
 import MockAdapter from 'axios-mock-adapter';
-import getLicenses from '../../actions/licenseActions';
-import types from '../../actions/actionTypes';
+import { types, getLicenses } from '../../actions/licenseActions';
 
 const middlewares = [thunk];
 const mockStore = configureMockStore(middlewares);
